% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDL.L.R
\name{HDL.L}
\alias{HDL.L}
\title{Local version of High-definition likelihood inference of genetic correlations (HDL-L)}
\usage{
HDL.L(
  gwas1.df,
  gwas2.df,
  Trait1name,
  Trait2name,
  LD.path,
  bim.path,
  Nref = 335272,
  N0,
  chr,
  piece,
  output.file = "",
  eigen.cut = 0.99,
  intercept.output = FALSE,
  fill.missing.N = NULL,
  lim = exp(-18),
  alpha = 0.05
)
}
\arguments{
\item{gwas1.df}{A data frame including GWAS summary statistics of genetic variants for trait 1.
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS. If the GWAS is based on logistic regression, `b` should be the logarithm of OR (odds ratio) and `se` is the standard error of log(OR). Notice: SNPs with missing N will be removed.}

\item{gwas2.df}{A data frame including GWAS summary statistics of genetic variants for trait 2.
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS. If the GWAS is based on logistic regression, `b` should be the logarithm of OR (odds ratio) and `se` is the standard error of log(OR). Notice: SNPs with missing N will be removed.}

\item{Trait1name}{The trait name for gwas1.df}

\item{Trait2name}{The trait name for gwas2.df}

\item{LD.path}{Path to the directory where linkage disequilibrium (LD) information is stored. see note below.}

\item{bim.path}{Path to the directory where the decompressed bimfile.zip file is stored. see note below.}

\item{Nref}{Sample size of the reference sample where LD is computed. If the default UK Biobank reference sample is used, Nref = 335272}

\item{N0}{Number of individuals included in both cohorts.}

\item{chr}{The chromosome to which the region belongs.}

\item{piece}{The piece of the genome to which the region belongs. The whole genome is divided into 2,468 smaller, semi-independent blocks, each defined by LD calculated by Plink. The SNP information in each local region is included in the data set: "HDLL_LOC_snps.RData".}

\item{output.file}{Where the log and results should be written. If you do not specify a file, the log will be printed on the console.}

\item{eigen.cut}{Specifies which eigenvalues and eigenvectors in each LD score matrix should be used for HDL. The default value is 0.99. Users are allowed to specify a numeric value between 0 and 1 for eigen.cut.}

\item{intercept.output}{Logical, FALSE by default. Should the intercept terms be included in estimates.df?}

\item{fill.missing.N}{If NULL (default), the SNPs with missing N are removed. One of "median", "min" or "max" can be given so that the missing N will be filled accordingly. For example, "median" means the missing N are filled with the median N of the SNPs with available N.}

\item{lim}{Tolerance limitation, default lim = exp(-18).}

\item{alpha}{The significance level used to compute the cut-off value 'c' for the likelihood-based confidence interval (LCI). The cut-off 'c' is calculated using the formula: c = exp(-qchisq(1 - alpha, 1) / 2). For example, when alpha = 0.05, it corresponds to a 95 percent likelihood-based confidence interval.}
}
\value{
A dataframe is returned with:

\itemize{
\item{Trait1 }{The name of trait 1.}
\item{Trait2 }{The name of trait 2.}
\item{chr }{Which chromosome.}
\item{piece }{which piece of the chromosome}
\item{estimates.df }{A detailed matrix includes the estimates, P-value or likelihood-based intervals of heritability, genetic covariance and genetic correlation.}
\item{eigen.use }{The eigen.cut used in computation.}
}
}
\description{
The function returns the estimate and standard error of the genetic correlation between two traits based on GWAS summary statistics.
}
\note{
Users can download the precomputed eigenvalues and eigenvectors of LD correlation matrices for European ancestry population. The download link can be found at https://doi.org/10.5281/zenodo.14825987 (LDfile.zip, bimfile.zip)
These are the LD matrices and their eigen-decomposition from 335,272 genomic British UK Biobank individuals.
}
\references{
Li Y, Pawitan Y, Shen X (2020). An enhanced framework for local genetic correlation analysis.
}
\seealso{
HDL tutorial: https://github.com/zhenin/HDL
}
\author{
Yuying Li, Xia Shen
}
