% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDL.rg.R
\name{HDL.rg}
\alias{HDL.rg}
\title{High-definition likelihood inference of genetic correlations (HDL)}
\usage{
HDL.rg(gwas1.df, gwas2.df, LD.path, Nref = 336000, N0 = min(gwas1.df$N,
  gwas2.df$N), output.file = "")
}
\arguments{
\item{gwas1.df}{A data frame including GWAS summary statistics of genetic variants for trait 1. 
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS.}

\item{gwas2.df}{A data frame including GWAS summary statistics of genetic variants for trait 2. 
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS.}

\item{LD.path}{Path to the directory where linkage disequilibrium (LD) information is stored.}

\item{Nref}{Sample size of the reference sample where LD is computed. If the default UK Biobank reference sample is used, Nref = 336000.}

\item{N0}{Number of individuals included in both cohorts. However, the estimated genetic correlation is usually robust against misspecified N0. 
If not given, the default value is set to the minimum sample size across all SNPs in cohort 1 and cohort 2.}

\item{output.file}{In current version, this argument is only for the convinience of command line user.}
}
\value{
A list is returned with:
\itemize{
\item{rg }{The estimated genetic correlation.}
\item{rg.se }{The standard error of the estimated genetic correlation.}
\item{P }{P-value based on Wald test.}
}
}
\description{
The function returns the estimate and standard error of the genetic correlation between two traits based on GWAS summary statistics.
}
\note{
Users can download the eigenvalues and eigenvectors of LD correlation matrices from https://www.dropbox.com/sh/ai2o21gxklhlvs3/AABPD7nKv3nXcvmoQQ3cGh9Qa?dl=0. 
These are the LD matrices and their eigen-decomposition from 336,000 genomic British UK Biobank individuals. Two sets of reference panel are provided: 
1) 307,519 QCed UK Biobank Axiom Array SNPs. The size is about 7.5 GB after unzipping.
2) 1,029,876 QCed UK Biobank imputed SNPs. The size is about 31 GB after unzipping. Although it takes more time, using the imputed panel provides more accurate estimates of genetic correlations. 
Therefore if the GWAS includes most of the HapMap3 SNPs, then we recommend using the imputed reference panel.
}
\examples{
\dontrun{
## The GWAS summary statistics for birth weight 
data(gwas1.example)

## The GWAS summary statistics for type 2 diabetes
data(gwas2.example)

## The path to the directory where linkage disequilibrium (LD) information is stored.
LD.path <- "/Users/zhengning/Work/HDL/package/UKB_SVD_eigen90_extraction"

res.HDL <- HDL.rg(gwas1.example, gwas2.example, LD.path)
res.HDL
}
}
\references{
Ning Z, Pawitan Y, Shen X (2019). High-definition likelihood inference of genetic correlations 
across human complex traits. \emph{Submitted}.
}
\seealso{
HDL tutorial: https://github.com/zhenin/HDL
}
\author{
Zheng Ning
}
