% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDL.rg.R
\name{HDL.rg}
\alias{HDL.rg}
\title{High-definition likelihood inference of genetic correlations (HDL)}
\usage{
HDL.rg(gwas1.df, gwas2.df, LD.path, Nref = 336000, N0 = min(gwas1.df$N,
  gwas2.df$N), output.file = "")
}
\arguments{
\item{gwas1.df}{A data frame including GWAS summary statistics of genetic variants for trait 1. 
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS.}

\item{gwas2.df}{A data frame including GWAS summary statistics of genetic variants for trait 2. 
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS.}

\item{LD.path}{Path to the directory where linkage disequilibrium (LD) information is stored.}

\item{Nref}{Sample size of the reference sample where LD is computed. If the default UK Biobank reference sample is used, Nref = 336000.}

\item{N0}{Number of individuals included in both cohorts. However, the estimated genetic correlation is usually robust against misspecified N0. 
If not given, the default value is set to the minimum sample size across all SNPs in cohort 1 and cohort 2.}

\item{output.file}{The file where the log should be saved. If not given, the log will not be saved.}
}
\value{
A list is returned with:
\itemize{
\item{rg }{The estimation of genetic correlation.}
\item{rg.se }{The standard error of estimated genetic correlation.}
}
}
\description{
The function returns estimation and standard error of genetic correlation between two traits based on GWAS summary statistics.
}
\note{
Users can download the eigenvalues and eigenvectors of LD correlation matrices from link. 
These eigens of LD are from eigen-decomposition results based on 307,519 QCed UK Biobank Axiom Array SNPs. 
336,000 ethnically British individuals in UK Biobank are used to compute LD.
}
\examples{
\dontrun{
## The GWAS summary statistics for birth weight 
data(gwas1.example)

## The GWAS summary statistics for type 2 diabetes
data(gwas2.example)

## The path to the directory where linkage disequilibrium (LD) information is stored.
LD.path <- "/Users/zhengning/Work/HDL/package/UKB_SVD_eigen90_extraction"

res.HDL <- HDL.rg(gwas1.example, gwas2.example, LD.path)
res.HDL
}
}
\references{
Ning Z, Pawitan Y, Shen X (2019). High-definition likelihood inference of genetic correlations 
across human complex traits. \emph{Submitted}.
}
\seealso{
HDL tutorial: https://github.com/zhenin/HDL
}
\author{
Zheng Ning
}
